Port my "ext_m_supplement.c" from my patch to the Linux kernel back to this project.  Here it is called "gcc_supp.c".
I have ported the preprocessor directives as well as the coding
style from the Linux kernel to this project.  Otherwise, it is the
same code.

This version should be more inline with correct and portable C, as
well as includes implementaions for the __yxxxti3() series of
integer arithmetic routines.  This should allow this code to work
on a possible future 64-bit port of this project.