Removed the hack from the Makefile
In a previous commit, I broke the Makefile by forcing it to clean
every time it was used to build.  This isn't such a a huge issue
now as the build time is very short, but it was certainly a nasty
hack.  The reason for this was that I was trying to quickly solve
the problem of 2 different configurations of this firmware being
mutally incompatible across 2 different emulators.  Preprocessor
definitions are currently used to control the configurations built.
I was trying to debug another issue and didn't have time to do a
proper build system at that time.

I have now changed it so that each configuration is a seperate
build target.  This still abuses GNU Make in a bit of a hacky way
because it makes use of some internal makefile targets that could
break if called directly.  However, it is better than the former.

A better way to handle this is might be to have a seperate
configuration step before make is used.  Perhaps something like
the Kbuild/Kconfig system used in the Linux Kernel, Busybox, and
Toybox.  For now, this is probably good enough.

I still need to make a few more tweeks to reintroduce the rebuild
target, etc...

I also removed the long unused bare-metal.ld linker script.  It
can be added back later if a use comes up but for now it is just
adding clutter.